pyproject.toml
[tool.poetry]
name = "doc-table"
version = "0.1.0"
description = ""
authors = ["遥奕 <jibing.yjb@alibaba-inc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyfunvice = "^0.1.5"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
